requirements.txt
dnspython==2.1.0
Flask==1.1.2
Flask-PyMongo==2.3.0
gunicorn==20.0.4
numpy==1.19.5
pymongo==3.11.3
python-dotenv==0.16.0
scikit-learn==0.24.1
tensorflow==2.4.1